requirements.txt
Flask==3.0.0
firebase-admin==6.5.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
requests==2.31.0
cryptography==42.0.5
itsdangerous==2.2.0
Werkzeug==3.0.1
python-jose[cryptography]==3.3.0
